fix(dm): drop ConversationScreen under size cap + share local wallet alias
Extract the shared-contact kind-0 profile fetch out of ConversationScreen
into useSharedContactProfiles, bringing the screen back under the 1,000-line
cap (1004 → 958) after the origin/main merge. The hook owns the sharedProfiles
state, the scheduled-pubkeys ref, and the batch-fetch effect — one nameable
responsibility, independently testable.

Per the Copilot review on useNwcShareActions, send the user's own local wallet
`alias` (the exact label the confirm dialog shows the sender) instead of the
remote getInfo `walletAlias`, falling back to walletAlias only when the local
label is blank — so the recipient's redacted share card matches what the sender
agreed to share.
