+++ src/screens/ConversationScreen.tsx
@@ -50,12 +50,11 @@ import { buildOsmViewUrl, SharedLocation } from '../services/locationService';
 import { useLiveLocation } from '../contexts/LiveLocationContext';
 import { useUserLocation } from '../contexts/UserLocationContext';
 import LiveLocationDurationPicker from '../components/LiveLocationDurationPicker';
-import { fetchProfile, DEFAULT_RELAYS } from '../services/nostrService';
 import { isConfigured as isGifConfigured } from '../services/giphyService';
 import type { NostrProfile } from '../types/nostr';
 import type { RootStackParamList } from '../navigation/types';
-import { extractSharedContact } from '../utils/messageContent';
 import { useNwcShareActions } from '../hooks/useNwcShareActions';
+import { useSharedContactProfiles } from '../hooks/useSharedContactProfiles';
 import { useConversationPolls } from '../hooks/useConversationPolls';
 import { isSupportedImageUrl } from '../utils/imageUrl';
 import { usePaidInvoiceTracker } from '../hooks/usePaidInvoiceTracker';
@@ -142,15 +141,10 @@ const ConversationScreen: React.FC = () => {
   const [invoiceToPay, setInvoiceToPay] = useState<string | null>(null);
   const [avatarError, setAvatarError] = useState(false);
   const [detailTx, setDetailTx] = useState<TransactionDetailData | null>(null);
-  // Profiles resolved from `nostr:` contact references the other party
-  // has shared in this conversation. Keyed by hex pubkey; a `null` value
-  // means we tried and the kind-0 lookup came back empty.
-  const [sharedProfiles, setSharedProfiles] = useState<Record<string, NostrProfile | null>>({});
-  // Tracks which pubkeys have already been scheduled for a kind-0 fetch
-  // so the batch-fetch effect deps can be [messages] only, without needing
-  // sharedProfiles in the array (which would cause an extra cycle after
-  // every fetch batch writes the state).
-  const scheduledProfilePubkeys = useRef(new Set<string>());
+  // Profiles resolved from `nostr:` contact references the other party has
+  // shared in this conversation — see useSharedContactProfiles. Keyed by hex
+  // pubkey; a `null` value means the kind-0 lookup ran and came back empty.
+  const sharedProfiles = useSharedContactProfiles(messages);
   const [attachPanelOpen, setAttachPanelOpen] = useState(false);
   // Secret Mode chat-trigger card overlay state — driven by
   // MessageBubble's "secretthreewords" magic message. Owned here so
@@ -270,46 +264,6 @@ const ConversationScreen: React.FC = () => {
   }, [items.length]);
 
   const { isInvoicePaid } = usePaidInvoiceTracker(messages);
-
-  // Batch-fetch profiles for every `nostr:` profile reference that appears
-  // in the conversation. Relay hints from the nprofile (when present) are
-  // merged with the default set so we find the shared person's kind-0
-  // even if they publish on niche relays.
-  useEffect(() => {
-    const byPubkey = new Map<string, Set<string>>();
-    for (const m of messages) {
-      const ref = extractSharedContact(m.text);
-      if (!ref) continue;
-      if (scheduledProfilePubkeys.current.has(ref.pubkey)) continue;
-      const set = byPubkey.get(ref.pubkey) ?? new Set<string>();
-      for (const r of ref.relays) set.add(r);
-      byPubkey.set(ref.pubkey, set);
-    }
-    if (byPubkey.size === 0) return;
-    // Mark all found pubkeys as scheduled before the async work starts so
-    // a second messages-update doesn't re-queue the same fetches.
-    for (const pk of byPubkey.keys()) scheduledProfilePubkeys.current.add(pk);
-    let cancelled = false;
-    (async () => {
-      const updates: Record<string, NostrProfile | null> = {};
-      await Promise.all(
-        [...byPubkey.entries()].map(async ([pk, relaySet]) => {
-          const mergedRelays = [...new Set([...DEFAULT_RELAYS, ...relaySet])];
-          try {
-            updates[pk] = await fetchProfile(pk, mergedRelays);
-          } catch {
-            updates[pk] = null;
-          }
-        }),
-      );
-      if (!cancelled) {
-        setSharedProfiles((prev) => ({ ...prev, ...updates }));
-      }
-    })();
-    return () => {
-      cancelled = true;
-    };
-  }, [messages]);
 
   // Contact preview sheet — peek a counterparty without leaving the
   // conversation. Tapping "View full profile" inside the sheet drills
